step2-cache_sink_file_detector/requirements.txt
bashlex==0.18
PyYAML==6.0.1
